requirements.txt
numpy==1.23.4
Pillow==9.2.0
streamlit==1.13.0
tflite_runtime==2.5.0.post1
